TOT-96 : [REFACTOR] Dto클래스 변경에 따른 테스트 코드 import문 수정
